requirements.txt
# Data Processing
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Statistical Analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0
imbalanced-learn>=0.11.0

# Model Interpretability
shap>=0.43.0
eli5>=0.13.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Web (existing)
Flask==3.0.3
